bugfix: improve cgroup v2 detection
The existing behavior checks for the presence of a
`/sys/fs/cgroup/cpu.max` file to determine if a filesystem is cgroup v2
or not. Unfortunately there is no guarantee that this file will exist in
a cgroup v2 filesystem. Instead we call the `statfs` syscall to
determine what type the filesystem is, checking it against the special
cgroup v2 magic number.