fix: correctly detect cgroup version
Fixes a bug where we always hit the fallback logic for cgroupv2
detection.

By default, `Statter` is created with an `afero.ReadOnlyFs`, which means
our `if _, ok := s.fs.(*afero.OsFs); ok` check completely falls apart.
This was only implemented like this as a way to allow us to use an
`afero.Fs` for testing. We instead ditch this logic entirely, and
introduce a field on `Statter` that handles the detection. We set
a default value for this that works on the host file system, and allow
setting a custom detector with `WithCGroupV2Detector`. This way we can
have control over which flow is hit during tests, without having
a brittle cgroup detection logic.